3.1.12 przeorganizowanie projektu - usunięcie pliku zawierająca klasę Node
